mark69/assistant/requirements.txt
# Raspberry Pi 5 Offline Multimodal AI Assistant
# Python 3.9+ recommended

# Display - OLED SSD1306
luma.oled>=3.8.0
luma.core>=2.4.0
Pillow>=10.0.0

# GPIO
RPi.GPIO>=0.7.1

# Camera
picamera2>=0.3.12

# Vision - YOLO
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Audio - STT (offline)
vosk>=0.3.45
pyaudio>=0.2.14
webrtcvad>=2.0.10

# Audio - TTS (offline)
piper-tts>=1.2.0
# Alternative: pyttsx3>=2.90 (optional, fallback)

# LLM - llama.cpp HTTP client
requests>=2.31.0
sseclient-py>=1.8.0

# Memory / RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
